+++ astropy/units/astrophys.py
@@ -46,24 +46,6 @@ def_unit(['lyr', 'lightyear'], (_si.c * si.yr).to(si.m),
 
 
 ###########################################################################
-# AREAS
-
-def_unit(['barn', 'barn'], 10 ** -28 * si.m ** 2, namespace=_ns, prefixes=True,
-         doc="barn: unit of area used in HEP")
-
-
-###########################################################################
-# ANGULAR MEASUREMENTS
-
-def_unit(['cycle', 'cy'], 2.0 * _numpy.pi * si.rad,
-         namespace=_ns, prefixes=False,
-         doc="cycle: angular measurement, a full turn or rotation")
-
-def_unit(['spat', 'sp'], 4.0 * _numpy.pi * si.sr,
-         namespace=_ns, prefixes=False,
-         doc="spat: the solid angle of the sphere, 4pi sr")
-
-###########################################################################
 # MASS
 
 def_unit(['solMass', 'M_sun', 'Msun'], _si.M_sun, namespace=_ns,
@@ -77,14 +59,6 @@ def_unit(['earthMass', 'M_earth', 'Mearth'], _si.M_earth, namespace=_ns,
          prefixes=False, doc="Earth mass",
          # LaTeX earth symbol requires wasysym
          format={'latex': r'M_{\oplus}', 'unicode': 'M⊕'})
-def_unit(['M_p'], _si.m_p, namespace=_ns, doc="Proton mass",
-         format={'latex': r'M_{p}', 'unicode': 'Mₚ'})
-def_unit(['M_e'], _si.m_e, namespace=_ns, doc="Electron mass",
-         format={'latex': r'M_{e}', 'unicode': 'Mₑ'})
-# Unified atomic mass unit
-def_unit(['u', 'Da', 'Dalton'], _si.u, namespace=_ns,
-         prefixes=True, exclude_prefixes=['a', 'da'],
-         doc="Unified atomic mass unit")
 
 ##########################################################################
 # ENERGY
@@ -99,13 +73,6 @@ def_unit(['Ry', 'rydberg'],
          doc="Rydberg: Energy of a photon whose wavenumber is the Rydberg "
          "constant",
          format={'latex': r'R_{\infty}', 'unicode': 'R∞'})
-
-##########################################################################
-# PRESSURE
-
-def_unit(['bar'], 1e5 * si.Pa, namespace=_ns,
-         prefixes=[(['m'], ['milli'], 1.e-3)],
-         doc="bar: pressure")
 
 ###########################################################################
 # ILLUMINATION
@@ -147,25 +114,12 @@ def_unit(['Sun'], namespace=_ns)
 def_unit((['ct', 'count'], ['count']),
          format={'fits': 'count', 'ogip': 'count', 'vounit': 'count'},
          namespace=_ns, prefixes=True, exclude_prefixes=['p'])
-def_unit((['pix', 'pixel'], ['pixel']),
-         format={'ogip': 'pixel', 'vounit': 'pixel'},
-         namespace=_ns, prefixes=True)
-
 
 ###########################################################################
 # MISCELLANEOUS
 
 def_unit(['chan'], namespace=_ns, prefixes=True)
 def_unit(['bin'], namespace=_ns, prefixes=True)
-def_unit((['vox', 'voxel'], ['voxel']),
-         format={'fits': 'voxel', 'ogip': 'voxel', 'vounit': 'voxel'},
-         namespace=_ns, prefixes=True)
-def_unit((['bit', 'b'], ['bit']), namespace=_ns,
-         prefixes=si_prefixes + binary_prefixes)
-def_unit((['byte', 'B'], ['byte']), 8 * bit, namespace=_ns,
-         format={'vounit': 'byte'},
-         prefixes=si_prefixes + binary_prefixes,
-         exclude_prefixes=['d'])
 def_unit(['adu'], namespace=_ns, prefixes=True)
 def_unit(['beam'], namespace=_ns, prefixes=True)
 def_unit(['electron'], doc="Number of electrons", namespace=_ns,
@@ -178,15 +132,6 @@ def_unit(['electron'], doc="Number of electrons", namespace=_ns,
 def_unit(['littleh'], namespace=_ns, prefixes=False,
          doc="Reduced/\"dimensionless\" Hubble constant",
          format={'latex': r'h_{100}'})
-
-# The torr is almost the same as mmHg but not quite.
-# See https://en.wikipedia.org/wiki/Torr
-# Define the unit here despite it not being an astrophysical unit.
-# It may be moved if more similar units are created later.
-def_unit(['Torr', 'torr'], _si.atm.value/760. * si.Pa, namespace=_ns,
-         prefixes=[(['m'], ['milli'], 1.e-3)],
-         doc="Unit of pressure based on an absolute scale, now defined as "
-             "exactly 1/760 of a standard atmosphere")
 
 ###########################################################################
 # CLEANUP